requirements.txt
pyserial==3.5
XlsxWriter==3.1.9
tk-tools==0.15.0
